fix(gemini): tighten Google tool loop
