--- src/core/providers/google/mod.rs
@@ -1,3 +1,0 @@
-//! Provider-neutral Google chat semantics shared by Gemini and Vertex AI.
-
-pub(crate) mod tool_loop;
